rezolvat cazuri speciale pbinfo 2699 2377

+++ Olimpiada/2377/main.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-    int n,m;
+    long long int n,m;
     cin>>n>>m;
     if(m==0)
         cout<<0;
